requirements.txt
pillow==11.0.0
Django==5.1.3
celery==5.4.0
redis==5.2.0
